Change container exit to int and add success exit code

+++ src/main/java/com/iexec/common/docker/DockerRunResponse.java
@@ -29,7 +29,7 @@ public class DockerRunResponse {
 
     private boolean isSuccessful;
     private DockerLogs dockerLogs;
-    private Long containerExitCode;
+    private int containerExitCode;
 
     public String getStdout() {
         if (dockerLogs != null && dockerLogs.getStdout() != null) {
